refactor: Arquivos fonte movidos para 'src'
