+++ src/components/ui/form.tsx
@@ -10,8 +10,8 @@ import {
   useFormContext,
 } from "react-hook-form"
 
-import { cn } from "@/lib/utils"
-import { Label } from "@/components/ui/label"
+import { cn } from "@/src/lib/utils"
+import { Label } from "@/src/components/ui/label"
 
 const Form = FormProvider
